Telecomm: Fix createLaunchEmergencyDialerIntent() crash in non-main users
TelecomServiceImpl and PackageManagerService both run inside the system_server, which means that
caller checks that are performed by mPackageManager.resolveActivity() are made against caller of
createLaunchEmergencyDialerIntent() instead of against TelecomServiceImpl itself.

This leads to an implicit and unnecessary requirement for callers to hold the privileged
INTERACT_ACROSS_USERS permission if they are running in a non-main user (mPackageManager is derived
from the context of the main user).

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Romeo <rbesar210311@gmail.com>